artifact_backend/requirements.txt
Django
python-dotenv
mysqlclient
djangorestframework
django-cors-headers